+++ model/__init__.py
@@ -2,7 +2,7 @@
     We are not just developers, we are artists and we are building the world of tomorrow
     -- Emmanuel ABO --
 """
-print("\t\t-----------------------------------------------\n\t\t|\t\t\t- PUF Simulation -                "
+print("\t\t-----------------------------------------------\n\t\t|\t\t\t- ArbiterPUF Simulation -                "
       "|\n\t\t|\tAuthors: "
       "ABO Emmanuel - KAMIDJIGHA "
       "Mounya |\n\t\t-----------------------------------------------")
